--- medimages4tests/mri/neuro/base.py
@@ -1,10 +1,0 @@
-from pathlib import Path
-import attrs
-
-
-@attrs.define
-class OpenneuroSpec:
-
-    dataset: str
-    tag: str
-    path: Path
